requirements.txt
fastapi
uvicorn[standard]
python-lorem
pytest
httpx
pytest-playwright
requests